requirements.txt
fastapi
uvicorn[standard]
motor
dnspython==2.3
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pydantic-settings
pydantic[email]
python-dotenv
python-multipart

langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-experimental==0.0.42
langchain-google-genai==3.0.0
google-ai-generativelanguage==0.9.0
google-api-core==2.27.0
google-auth==2.41.1
googleapis-common-protos==1.71.0

tiktoken
faiss-cpu
youtube-transcript-api
numpy

PyPDF2
python-docx